pyproject.toml
[tool.poetry]
name = "fast-api-learning"
version = "0.1.0"
description = ""
authors = ["Alexander Scherbakov <webmasterdracarys@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
requests = "^2.32.3"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
email-validator = "^2.2.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.17"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
flower = "^2.0.1"
pillow = "^11.0.0"
celery = "^5.4.0"
jinja2 = "^3.1.4"
aiofiles = "^24.1.0"

[tool.poetry.group.dev.dependencies]
fastapi = "^0.115.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
